requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyMySQL==1.0.2
python-dotenv==0.21.1
SQLAlchemy==1.4.46
Werkzeug==2.2.2
zipp==3.11.0
